Consolidate native methods into Win32Sec
